Moved QueryParamBuilder into individual component.
- Fixed issues around `inProgress` flags for tabs.
