+++ src/renderer/store/middlewares/persistor.ts
@@ -15,7 +15,7 @@ const persistMiddleware: ReduxMiddlewareType = ({ getState }) => {
       layout: {
         ...currentState.layout,
         notifications: [],
-        tabs: currentState.layout.tabs.map(tab => omit(tab, ['meta', 'results'])),
+        tabs: currentState.layout.tabs.map(tab => omit(tab, ['inProgress', 'meta', 'results'])),
       },
     };
